+++ src/components/ContactCard/ContactCard.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { translate } from 'cozy-ui/transpiled/react/I18n'
 import ContactIdentity from './ContactIdentity'
-import ContactGroupsList from '../ContactGroups/ContactGroupList'
+import ContactGroupsList from './ContactGroups/ContactGroupList'
 import ContactFields from './ContactFields/ContactFields'
 import { fullContactPropTypes } from '../ContactPropTypes'
 import { normalizeFields, getConnectedAccounts } from '../../helpers/contacts'
