+++ src/components/ContactCard/ContactGroups/ContactGroupList.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import { fullContactPropTypes } from '../ContactPropTypes'
-import FieldsetTitle from '../Common/FieldsetTitle'
-import GroupItem from '../Common/GroupItem'
+import { fullContactPropTypes } from '../../ContactPropTypes'
+import FieldsetTitle from '../../Common/FieldsetTitle'
+import GroupItem from '../../Common/GroupItem'
 
 export const ContactGroupsList = ({ contact, allGroups, title }) => {
   const userGroups = contact.groups.data
