+++ src/components/GroupsSelect/SelectBox/Option.jsx
